build: quality of life improvements

+++ setup.sh
@@ -1,12 +1,17 @@
 #!/bin/bash
 
-# set this yourself
-export PICO_SDK_PATH=/Users/stuart/Documents/builds/pico/pico-sdk
-export PICO_TOOLCHAIN_PATH=/Applications/ARM/bin
+export GBCAM_PATH="$(cd "$(dirname "$0")" && pwd)"
+
+if ! [ -f "$GBCAM_PATH/.config.local.sh" ]; then
+    echo "Error: Must create .config.local.sh based off .config.sh" 1>&2
+    return 1
+fi
+
+source "$GBCAM_PATH/.config.local.sh"
 
 (
     set -e
-    cd "$(dirname "$0")"
+    cd "$GBCAM_PATH"
 
     rm -rf build/
     mkdir build
